Squashed commit of loss accounting and reporting feature.
Original commits in PR #21.
Add loss accounting logic to genotype_sites with accompanying data structure and tests. Loss is computed based on no-calls in the output vcf and loss attributed to original gvcf records is reported in addition to overall loss. Loss is conveyed to the users via stderr printing in CLI's main genotyping service.