In translate_genotypes, check for changes in REF allele.
In cases where the REF allele in the unified site differs from the original input (g)vcf, the ALT allele, when taken at face value, may have different meanings. This commit checks for changes in the REF allele and in the cases where the REF differs, all ALT calls in the original (g)vcf is deemed to be untranslatable and are genotyped as no-calls.