Join records in genotyper - check span & ref depth
Completed logic for joining records in genotyper v1. Defined notion of "representative record" for a site where multiple records are found.

Added check to assert that records span the entire range of site, and that sufficient depth is observed in the gvcf_ref records that are used for "joining".

TODO: Add documentation on notion of "representative record"