Genotyper v1 - join records / find representative records.
This is a squash merge for genotyper v1 feature. The complete commit history can be found in the corresponding PR (#55).

Genotyper v1 is an improvement upon the previous placeholder genotyper algorithm. In particular, it makes 2 main change, which are predicate upon unifier v1 implementation:

1. Allow unification lookup (in translate_genotype) for records that do not span the exact range as the unified_site; relax REF lookup to any record that overlaps unified_site.

2. Manage unified_sites that has multiple records overlapping with it. We handle the following 2 cases:
  - When all records are gvcf confidence records, we interpret the site as REF
  - When only 1 record is a variant record, we take this as the "representative record" used fot translate_genotype.

Added additional gvcf_test_cases tests (join_records*.yml) for testing of the above changes, and to support filtering of depth in cases where reference padding is implicated.