BCFKeyValueData: reduce bucket scanning by duplicating records that cross two buckets
If a record dangles from one bucket to the next it's copied in both buckets (which each remain beg-sorted). Therefore, to query a range within a bucket, it's no longer necessary to scan the preceding bucket as well. However, if the query is such that we do need to scan multiple adjacent buckets, we will see dangling records twice, so we need to de-duplicate them (by ignoring danglers in all but the first scanned bucket).